feat: :sparkles: fetch関数の型定義
